refactor: move q8 file read telemetry helpers
